Move Validationslabel to modals subdirectory
