+++ ui/client/common/Validators.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 export const canNotBeEmpty = "This field cannot be empty"
-export const duplicate = (value) => value + " already taken"
+export const duplicateValue = "This value is already taken"
 
 export const notEmptyValidator = {
   isValid: (value) => !_.isEmpty(value),
